# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  diprec_api_internal_domain.CourseResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  diprec_api_internal_domain.CourseResponseWithTests:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      tests:
        items:
          $ref: '#/definitions/diprec_api_internal_domain.TestResponse'
        type: array
      updatedAt:
        type: string
    type: object
  diprec_api_internal_domain.Error:
    properties:
      message:
        type: string
    type: object
  diprec_api_internal_domain.QuestionAnswer:
    properties:
      answer: {}
      isCorrect:
        type: boolean
      message:
        type: string
    type: object
  diprec_api_internal_domain.QuestionResponse:
    properties:
      answer: {}
      id:
        type: integer
      title:
        type: string
      type:
        enum:
        - SINGLE
        - MULTIPLE
        - TEXT
        - NUMBER
        example: SINGLE
        type: string
      variants:
        additionalProperties: true
        type: object
    type: object
  diprec_api_internal_domain.TestResponse:
    properties:
      assignee:
        type: string
      createdAt:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  internal_transport_http_course.CreateCourseDTO:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  internal_transport_http_course.UpdateCourseDTO:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  internal_transport_http_question.CheckAnswerDTO:
    properties:
      answer: {}
    type: object
  internal_transport_http_question.CreateQuestionDTO:
    properties:
      answer: {}
      title:
        type: string
      type:
        enum:
        - SINGLE
        - TEXT
        - NUMBER
        - MULTIPLE
        example: SINGLE
        type: string
      variants:
        additionalProperties: true
        type: object
    type: object
  internal_transport_http_question.UpdateQuestionDTO:
    properties:
      answer: {}
      title:
        type: string
      type:
        enum:
        - SINGLE
        - TEXT
        - NUMBER
        - MULTIPLE
        example: SINGLE
        type: string
      variants:
        additionalProperties: true
        type: object
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для дипломного проекта
  termsOfService: http://swagger.io/terms/
  version: "1.0"
paths:
...
  /course:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.CourseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Получить курсы
      tags:
      - Course
    post:
      parameters:
      - description: Название и описание курса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_course.CreateCourseDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.CourseResponse'
      security:
      - BearerAuth: []
      summary: Создать курс
      tags:
      - Course
  /course/{id}:
    delete:
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Курс успешно удалён
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Удалить курс
      tags:
      - Course
    get:
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.CourseResponseWithTests'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Получить курс по ID
      tags:
      - Course
    put:
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      - description: Название и описание курса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_course.UpdateCourseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.CourseResponseWithTests'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Обновить курс
      tags:
      - Course
  /course/{id}/enroll:
    post:
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Записаться на курс
      tags:
      - Course
  /question:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/diprec_api_internal_domain.QuestionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Получить все вопросы
      tags:
      - Question
    post:
      parameters:
      - description: ДТО создания вопроса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_question.CreateQuestionDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.QuestionResponse'
      security:
      - BearerAuth: []
      summary: Создать вопрос
      tags:
      - Question
  /question/{id}:
    delete:
      parameters:
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Удалить вопрос
      tags:
      - Question
    get:
      parameters:
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Получить вопрос по ID
      tags:
      - Question
    put:
      parameters:
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      - description: ДТО обновления вопроса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_question.UpdateQuestionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Обновить вопрос
      tags:
      - Question
  /question/{id}/check:
    post:
      parameters:
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      - description: ДТО ответа на вопрос
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_question.CheckAnswerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.QuestionAnswer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Проверить вопрос
      tags:
      - Question
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"